Refactor token auth implementation to extend to gRPC
