--- frontend/server/src/main/java/org/pytorch/serve/util/TokenType.java
@@ -1,7 +1,0 @@
-package org.pytorch.serve.util;
-
-public enum TokenType {
-    INFERENCE,
-    MANAGEMENT,
-    TOKEN_API
-}
